Updated Adenita to work in SAMSON 0.8.0 Updated UI to make it suited to vertical docking minor readme update

+++ AdenitaCoreSE/modules/DAS/sources/DASDaedalus.cpp
@@ -950,7 +950,7 @@ void DASDaedalus::SetVerticesPositions(ADNPointer<ADNPart> origami, DASPolyhedro
     v = vertices.at(nId);
     ++count;
   }
-  cm /= vertices.size();
+  cm /= boost::numeric_cast<int>(vertices.size());
   // center
   if (center) {
     for (auto vit = vertices.begin(); vit != vertices.end(); ++vit) {
